Main Menu, Level Select, Pause Menu
The main menu can go to the level select, and the level select can go back to the main menu. The level select can go the level 1 prototype, but because there is no pause button there yet, the pause menu cannot be accessed. The pause menu however should be able to go back to the level or quit to the menu. The pause menu will not pause the level, rather it is just a scene for now that can maneuver between other scenes.